requirements.txt
Django==5.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
   